Avoid bashisms in the call to acme.sh
This command is using features that are not POSIX-compliant (the [[ and
the || in them), that despite being supported by many shells (including
bash) are not available in dash.

Since system() runs the code with /bin/sh that is an interface defined
as only being POSIX-compliant, with no extensions, it should not presume
a specific shell in that.

Closes: #6326
Signed-off-by: Mattia Rizzolo <mattia@mapreri.org>